% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{add_q_marks}
\alias{add_q_marks}
\title{Add quotation marks to value if necessary}
\usage{
add_q_marks(value)
}
\arguments{
\item{value}{Value of a variable to be enquoted}
}
\description{
Add quotation marks to value if necessary
}
