% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ano_donations_2017}
\alias{ano_donations_2017}
\title{Donations received by ANO in the campaign for the parliamentary election in 2017}
\format{A data frame with 161 rows and 8 variables:
\describe{
  \item{datum}{Date}
  \item{částka Kč}{Amount of donation in CZK}
  \item{příjmení}{Last name of the candidate}
  \item{jméno}{First name of the candidate}
  \item{titul před}{Academic titles written before the name}
  \item{titul za}{Academic titles written after the name}
  \item{datum narození}{Date of birth}
  \item{obec}{Place of residence of the donor (name)}
  ...
}}
\source{
\url{https://udhpsh.cz/}
}
\usage{
ano_donations_2017
}
\description{
Donations received by ANO in the campaign for the parliamentary election in 2017
}
\keyword{datasets}
