% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{calculate_similarity_between_persons}
\alias{calculate_similarity_between_persons}
\title{Calculate similarity between persons}
\usage{
calculate_similarity_between_persons(original, similar)
}
\arguments{
\item{original}{Data.frame with all data about original person}

\item{similar}{Data.frame with all data about similar persons
Calculate in how many columns the data about the persons match (strictly equals)
TODO: use other comparison than strict equality}
}
\description{
Calculate similarity between persons
}
