% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_panel_output}
\alias{create_panel_output}
\title{Create tidy output}
\usage{
create_panel_output(output, id = "id")
}
\arguments{
\item{output}{Data.frame resulting from find_all_similar}
}
\description{
Create tidy output
}
