% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_query_filter}
\alias{create_query_filter}
\title{Create filter for querying the DB (WHERE clauses)}
\usage{
create_query_filter(query_var)
}
\arguments{
\item{query_var}{Variables to be queried
(list of column, value, tolerance and operation to be done)}
}
\description{
Create filter for querying the DB (WHERE clauses)
}
