% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{create_var_list}
\alias{create_var_list}
\title{create list storing column with attributes for comparison,
operation on which the data should be compared (=, >, <, >=, <=)}
\usage{
create_var_list(x, operation)
}
\arguments{
\item{x}{Vector of variables}

\item{operation}{Operation sign (=, >, <, >=, <=)}
}
\description{
create list storing column with attributes for comparison,
operation on which the data should be compared (=, >, <, >=, <=)
}
