% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{find_missing}
\alias{find_missing}
\title{Find IDs of entitites from target that does not occur in output}
\usage{
find_missing(target, target_ids, found_ids)
}
\arguments{
\item{target}{Data.frame with target data}

\item{target_ids}{Column name of IDs in target data.frame}

\item{found_ids}{Vector with IDs found by find_all_similar function(s)}
}
\description{
Find IDs of entitites from target that does not occur in output
}
