% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{find_similar}
\alias{find_similar}
\title{Find match between one person from the first dataset (t1) and
the whole second dataset (t2)}
\usage{
find_similar(source, target, row, eq = NULL, eq_tol = NULL, eq_sub = NULL,
  ht = NULL, lt = NULL, hte = NULL, lte = NULL, id = "row_id",
  compare_cols = NULL, keep_duplicities = TRUE, verbose = TRUE)
}
\arguments{
\item{source}{Name of the first table}

\item{target}{Name of the second table}

\item{row}{Row from the first table for which similar person is searched}

\item{eq}{Vector of variables which should be equal in t1 and t2}

\item{eq_tol}{Vector of variables which should be approximately equal
in t1 and t2 within specified range (tolerance)}

\item{eq_sub}{Vector of variables in which value of t1 should be a substring of t2
(optimized for double last names after marriage)}

\item{ht}{Vector of variables which should be higher in t1}

\item{lt}{Vector of variables which should be lower in t1}

\item{hte}{Vector of variables which should be higher or equal in t1}

\item{lte}{Vector of variables which should be lower or equal in t1}

\item{id}{Column in source and target datasets containing ID of a row}

\item{compare_cols}{Columns to be used for comparison to remove duplicates}

\item{keep_duplicities}{Bool indicating whether entities with the same attributes
should be kept or the most similar entity to the original record should be found
(and duplicities should be removed)}

\item{verbose}{Specify if you want to display message for every 250th row}
}
\description{
Find match between one person from the first dataset (t1) and
the whole second dataset (t2)
}
