% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{insert_nonconsecutive}
\alias{insert_nonconsecutive}
\title{Insert nonconsecutive entities into pivot table}
\usage{
insert_nonconsecutive(out, noncons_pivot, source, target)
}
\arguments{
\item{out}{Pivot table returned by find_all_similar}

\item{noncons_pivot}{Pivot table returned by find_all_similar
between nonconsecutive datasets}

\item{source}{Name of the nonconsecutive pivot source}

\item{target}{Name of the nonconsecutive pivot target}
}
\description{
Insert nonconsecutive entities into pivot table
}
