% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kraj_2012}
\alias{kraj_2012}
\title{Candidate list from the 2012 regional election in the Czech Republic}
\format{A data frame with 7725 rows and 17 variables:
\describe{
  \item{row_id}{serial ID of row}
  \item{first_name}{First name of the candidate}
  \item{last_name}{First name of the candidate}
  \item{titles}{Academic titles}
  \item{highest_titles}{the highest attained title}
  \item{gender}{}
  \item{approx_birth_year}{Calculated year of birth using
  the year of election minus age of the candidate}
  \item{occupation}{Occupation}
  \item{residence}{Place of residence of the candidate}
  \item{region}{Region}
  \item{nominate_party}{Political party which nominated the candidate}
  \item{member_party}{Political party which is the candidate member of}
  \item{party}{Name of the party}
  \item{abs_votes}{Preferential votes the candidate received}
  \item{pct_votes}{Percentage share of preferential votes}
  \item{rank}{Rank}
  \item{mandate}{Mandate}
  ...
}}
\source{
\url{https://volby.cz/}
}
\usage{
kraj_2012
}
\description{
A dataset containing candidate lists
}
\keyword{datasets}
