% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psp_2017}
\alias{psp_2017}
\title{Candidate list from the 2017 parliamentary election in the Czech Republic}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7540 rows and 20 columns.}
\source{
\url{https://volby.cz/}
}
\usage{
psp_2017
}
\description{
A dataset containing candidate lists of parties running in the election to
the Chamber of Deputies in 2017.
}
\keyword{datasets}
