% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{return_missing_data}
\alias{return_missing_data}
\title{Return data of entities which occur in target dataset but not in source dataset}
\usage{
return_missing_data(target, row_id, missing)
}
\arguments{
\item{target}{target dataset}

\item{row_id}{Column name of row IDs}

\item{missing}{result of find_missing function}
}
\description{
Return data of entities which occur in target dataset but not in source dataset
}
