% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{return_nonconsecutive_data}
\alias{return_nonconsecutive_data}
\title{Return data of entities that occur in the source dataset but not in the target}
\usage{
return_nonconsecutive_data(result, source, target, row_id)
}
\arguments{
\item{result}{result of find_all_similar function}

\item{source}{source dataset}

\item{target}{target dataset}

\item{row_id}{Column name of row IDs}
}
\description{
Return data of entities that occur in the source dataset but not in the target
}
