% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{upload_data}
\alias{upload_data}
\title{Append row_id and upload to DB}
\usage{
upload_data(connection, df, name)
}
\arguments{
\item{connection}{SQLite connection}

\item{df}{data.frame to be uploaded}

\item{name}{name of the table}
}
\description{
Append row_id and upload to DB
}
