% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{append_missing}
\alias{append_missing}
\title{Append missing entities to output}
\usage{
append_missing(target, target_ids, out)
}
\arguments{
\item{target}{data.frame with target data}

\item{target_ids}{column name of IDs in target data.frame}

\item{out}{output returned by find_all_similar function}
}
\description{
Append missing entities to output
}
