% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{find_all_similar}
\alias{find_all_similar}
\title{Find all matches between tables t1 and t2 and add candidates who did not run in the
sequential election with missing values}
\usage{
find_all_similar(conn, t1, t2, ...)
}
\arguments{
\item{conn}{Connection to SQL database in which the data are stored}

\item{t1}{Name of the first table}

\item{t2}{Name of the second table}

\item{...}{Vectors containing variables for comparison (see find_similar)}
}
\description{
Find all matches between tables t1 and t2 and add candidates who did not run in the
sequential election with missing values
}
