% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{find_missing}
\alias{find_missing}
\title{Find IDs of entitites from target that does not occur in output}
\usage{
find_missing(target, target_ids, found_ids)
}
\arguments{
\item{target}{data.frame with target data}

\item{target_ids}{column name of IDs in target data.frame}

\item{out}{output returned by find_all_similar function}
}
\description{
Find IDs of entitites from target that does not occur in output
}
